pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "RafaelBochi", email = "rafaelbochibarros@gmail.com"},
]
dependencies = [
    "django>=4.2.7",
    "django-cors-headers>=4.3.1",
    "setuptools>=69.0.2",
    "django-restframework>=0.0.1",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "drf-spectacular>=0.26.5",
    "gunicorn>=21.2.0",
    "whitenoise[brotli]>=6.6.0",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

